post앱의 Serializers, Views, models에 대해 모듈화 수행, member앱의 Serializers 와 Views에 대해 모듈화 수행
