+++ django_app/member/views/__init__.py
@@ -1,0 +1,4 @@
+from .register import *
+from .logout import *
+from .login import *
+from .userupdate import *
