+++ django_app/post/models/__init__.py
@@ -1,0 +1,5 @@
+from .post import *
+from .comment import *
+from .hashtag import *
+from .like_or_book import *
+from .img import *
