+++ django_app/post/models/hashtag.py
@@ -1,0 +1,10 @@
+from django.db import models
+
+__all__ = ('HashTag', )
+
+
+class HashTag(models.Model):
+    name = models.CharField(unique=True, max_length=20)
+
+    def __str__(self):
+        return self.name
