+++ django_app/post/serializers/__init__.py
@@ -1,0 +1,5 @@
+from .comment import *
+from .hashtag import *
+from .like_and_bookmark import *
+from .post_detail import *
+from .post_list import *
