+++ django_app/post/serializers/post_detail.py
@@ -1,37 +1,9 @@
 import random
-
 from rest_framework import serializers
+from post.models import Post, HashTag, DefaultImg
+from post.serializers import HashTagSerializer
 
-from post.models import Comment, DefaultImg
-from post.models import Post, HashTag, PostLike, PostBookMark
-
-
-class HashTagSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = HashTag
-        fields = ('name',)
-
-
-class CommentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Comment
-        fields = ('id', 'post', 'content', 'author', 'created_date')
-
-
-class PostListSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Post
-        fields = (
-            'id',
-            'content',
-            'author',
-            'created_date',
-            'modified_date',
-            'like_users_counts',
-            'comments_counts',
-            'distance',
-            'img_thumbnail',
-            )
+__all__ = ('PostDetailSerializer', 'PostCreateSerializer', )
 
 
 class PostDetailSerializer(serializers.ModelSerializer):
@@ -41,7 +13,6 @@ class PostDetailSerializer(serializers.ModelSerializer):
         model = Post
         fields = ('id', 'content', 'author', 'created_date', 'modified_date', 'view_counts',
                   'like_users_counts', 'distance','is_bookmarked', 'is_like','comments_counts', 'hashtags', 'img')
-
 
     def save(self):
         return super().save()
@@ -55,7 +26,7 @@ class PostDetailSerializer(serializers.ModelSerializer):
 
         post.save()
 
-        if hashtags!= None:
+        if hashtags != None:
             post.hashtags.all().delete()
             for hashtag in hashtags:
                 h, created = HashTag.objects.get_or_create(name=hashtag)
@@ -82,18 +53,3 @@ class PostCreateSerializer(serializers.ModelSerializer):
                 h, created = HashTag.objects.get_or_create(name=hashtag)
                 post.hashtags.add(h)
         return post
-
-
-class PostLikeSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = PostLike
-        fields = ('like_user', 'post')
-
-
-class PostBookMarkSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = PostBookMark
-        fields = ('bookmark_user', 'post')
-
