+++ django_app/post/views/__init__.py
@@ -1,0 +1,4 @@
+from .post_list import *
+from .comment import *
+from .like_and_book import *
+from .post_detail import *
